feat(bearings): 将极限转速字段类型从decimal改为int并添加商家英文名称支持
- 将轴承相关的极限转速字段(LimitingSpeed、LimitingSpeedGrease、LimitingSpeedOil)
  从decimal?类型更改为int?类型以提高数据精度匹配度

- 在商家实体中添加EnglishName字段及其对应的设置方法

- 在创建和更新商家命令中添加对EnglishName字段的支持

- 修复批量创建商家时的字段映射错误，将EnglishName正确设置到商家实体

- 从同步DTO中移除不再需要的SortOrder和Description等字段

+++ src/OpenFindBearings.Application/Commands/Bearings/CreateBearing/CreateBearingCommand.cs
@@ -112,17 +112,17 @@ namespace OpenFindBearings.Application.Commands.Bearings.CreateBearing
         /// <summary>
         /// 极限转速 (rpm)
         /// </summary>
-        public decimal? LimitingSpeed { get; init; }
+        public int? LimitingSpeed { get; init; }
 
         /// <summary>
         /// 脂极限转速 (rpm)
         /// </summary>
-        public decimal? LimitingSpeedGrease { get; init; }
+        public int? LimitingSpeedGrease { get; init; }
 
         /// <summary>
         /// 油极限转速 (rpm)
         /// </summary>
-        public decimal? LimitingSpeedOil { get; init; }
+        public int? LimitingSpeedOil { get; init; }
 
         /// <summary>
         /// 轴承类型ID
